src/test/java/se/wederbrand/advent_2017/Day07Test.java
package se.wederbrand.advent_2017;

import org.junit.Test;

import static org.junit.Assert.assertEquals;

public class Day07Test {

	public static final String INPUT = "jovejmr (40)\n" +
		"fesmk (24)\n" +
		"gwhfv (74)\n" +
		"vxfoyx (101) -> aqytxb, ltnnn\n" +
		"pvtnv (77)\n" +
		"cpmuhnf (39)\n" +
		"ocezven (77)\n" +
		"xjqta (42)\n" +
		"hkckef (21)\n" +
		"nwuhqn (63)\n" +
		"wbbfc (49)\n" +
		"oictp (50)\n" +
		"zmizid (85)\n" +
		"uukembw (1054) -> yowehs, ytoju, jwyhe, bxqlx\n" +
		"iqdna (212) -> gmhwcj, vllsfc, ebptuar, lmcqa\n" +
		"vakapy (132) -> eiisyk, nghznvw, dqslnjk\n" +
		"qouck (77)\n" +
		"pqjeof (12)\n" +
		"faszmg (53)\n" +
		"jvjlfb (56)\n" +
		"oxwpxqj (44)\n" +
		"ryonf (18)\n" +
		"jrgndow (70)\n" +
		"iayecc (86)\n" +
		"lwuwrp (84)\n" +
		"haxfzky (72)\n" +
		"ahbxz (12)\n" +
		"muuyp (21)\n" +
		"emjuzdq (46)\n" +
		"csiof (258) -> rjyyh, rdqisk\n" +
		"orkisb (88)\n" +
		"yheewtp (13)\n" +
		"utrhfs (10)\n" +
		"kczko (84)\n" +
		"iwcqa (98)\n" +
		"chdtaz (59)\n" +
		"pqoybi (74) -> bsmwi, hsycjdc\n" +
		"pinapwf (88)\n" +
		"vpfsdll (93)\n" +
		"ohjniy (139) -> tvqfs, uyfca\n" +
		"cbldohy (92)\n" +
		"vvgyhb (526) -> ruqgy, xrgsnsh, wytylnc\n" +
		"wwjqcpu (90)\n" +
		"xvjuf (79)\n" +
		"zdsyu (67) -> ecstxkl, paofmv\n" +
		"eovjzwr (55)\n" +
		"ohsjl (36)\n" +
		"eduzwi (49) -> segwmpm, bfekpxz, gewtd, knvukf\n" +
		"oyhvxt (5)\n" +
		"kyoypr (69)\n" +
		"mfaftsr (34)\n" +
		"heqou (193) -> uriolfn, gfhfmig, crxpsq\n" +
		"ktbxbnn (132) -> fhzzvtv, oytgp\n" +
		"yscxlz (40)\n" +
		"dukpbd (44)\n" +
		"kzlpr (6)\n" +
		"qswzv (86)\n" +
		"pxjuos (61)\n" +
		"mvavmml (53)\n" +
		"kqxpxlx (57)\n" +
		"ixdllwp (1853) -> sialf, kjvaqy\n" +
		"nhrapmw (51) -> pifqyxs, wxloqgs, jkcrf, bmivh, btjjjxw, gfyvv, tpdaf\n" +
		"wurmmqn (14)\n" +
		"dqjxyyd (53)\n" +
		"vlnnmg (94)\n" +
		"rjyyh (68)\n" +
		"hwezjo (43349) -> rpyji, uukembw, pbthb\n" +
		"iyeuk (74)\n" +
		"lcaeg (92)\n" +
		"gfhfmig (44)\n" +
		"iugub (226) -> rauprc, bummi\n" +
		"tfpsrke (70)\n" +
		"nhjbg (267) -> fviaonx, uemcawn\n" +
		"rleprim (45) -> lruhxda, qxfvj\n" +
		"yowehs (1257) -> ynhgsj, deboo, jrovat\n" +
		"lefvu (10) -> pzths, vygkoys, etzfe, sudep\n" +
		"kcjcpk (9)\n" +
		"zhvyfkx (56)\n" +
		"ixeev (77)\n" +
		"rkjfzl (241) -> jakjzay, znfjlz, zftvgdy, bzhzzv, dhktab\n" +
		"cstxcb (31)\n" +
		"bzhzzv (212)\n" +
		"sqkae (73)\n" +
		"rpkaem (14)\n" +
		"krepomq (36)\n" +
		"hxdrc (4240) -> fcoaxeb, geniumj, wprzdkf, ygbuuxe, fvqwcn, dketeda, ozmbhy\n" +
		"nebobeh (37) -> pinapwf, qijqu, kccdhut\n" +
		"jkcrf (177) -> dahvvo, tbyfd\n" +
		"jlnxpy (156) -> akmzxzw, yrrhqap\n" +
		"xdqma (201) -> ggsdtdp, rbxfyau\n" +
		"dtpzna (83)\n" +
		"wovwg (83)\n" +
		"eaggakx (50)\n" +
		"imfepc (5)\n" +
		"uruco (74)\n" +
		"qbhev (1169) -> dptakt, qrsyoj, gzttehr, qwcyeqz\n" +
		"bvrxeo (3248) -> jnyexah, ltleg, fdnmqri, iysgr, dffdie, vvgyhb\n" +
		"exhjhpu (9)\n" +
		"zcubrms (41) -> bgwlsu, grlbob, lcaeg, dmdgesa\n" +
		"bqayvhe (142) -> nnsaju, xobzucu, qrlyc\n" +
		"kutvex (178) -> xqscl, jowbyue\n" +
		"tzlsvpk (348) -> pqyjyl, yrudth\n" +
		"jbqnyve (70)\n" +
		"aexqgs (98) -> ighxxvd, orkisb\n" +
		"sbgug (49) -> qmbfra, jlnxpy, zkljkp, ikves\n" +
		"mmsesaw (98)\n" +
		"xsnbmt (98)\n" +
		"pzhayls (59)\n" +
		"iqyrubd (73)\n" +
		"nwzzgtf (71)\n" +
		"jgbce (79)\n" +
		"blhjjx (81)\n" +
		"xyoatzj (79)\n" +
		"drukt (87)\n" +
		"bxxpwa (12)\n" +
		"uyfca (46)\n" +
		"qslflr (37)\n" +
		"czrmixj (84)\n" +
		"rogrsp (94)\n" +
		"bcmep (79)\n" +
		"tvqfs (46)\n" +
		"vhdfs (50)\n" +
		"quaqp (29)\n" +
		"kqaox (70)\n" +
		"lyetyd (1427) -> craakno, unxgbyg, ftmwpqg\n" +
		"uktdx (135) -> oyhvxt, imfepc\n" +
		"eyhbm (45)\n" +
		"vrlwo (79) -> nbdhcy, adsppu, dmwzg, odguqiy\n" +
		"xeoeht (67)\n" +
		"szgqmow (110) -> oarrrpk, nqdcbhp\n" +
		"hmxbpwl (12)\n" +
		"xorzt (21)\n" +
		"nohqdkf (39)\n" +
		"fbshzlu (139) -> ahjpc, tgspt\n" +
		"zlrod (188)\n" +
		"ftudp (56)\n" +
		"zskbac (94) -> oyvsrsc, dmroo, wfrqy, bbpyckh\n" +
		"sdkbbc (94)\n" +
		"qngcb (40)\n" +
		"xziemca (23)\n" +
		"rpvqak (300) -> mfpxtom, cltcdp, khcau, bqayvhe, efoouqi\n" +
		"bojkbqr (75)\n" +
		"jsthfix (191) -> sewnmby, qngcb\n" +
		"lxckefh (21)\n" +
		"qrlyc (22)\n" +
		"aemfgyt (156)\n" +
		"xbegaua (106) -> cbcxlyo, cejbp\n" +
		"hfcotwm (36)\n" +
		"wrmngqs (9)\n" +
		"anpriwx (67) -> uhczfn, ovpmj\n" +
		"gsuxkc (66)\n" +
		"qryxc (74)\n" +
		"tymwxrt (67)\n" +
		"ifmnxwy (57)\n" +
		"qwjcmlq (5)\n" +
		"qlmptj (82)\n" +
		"eazsmg (75)\n" +
		"kygcd (58)\n" +
		"surfn (36)\n" +
		"ikwre (46) -> lpsufo, eaugk\n" +
		"tdulat (184) -> qfcknc, dbzmyy\n" +
		"yophj (74)\n" +
		"gverkkt (51)\n" +
		"vwhmf (54)\n" +
		"krvsaqc (51)\n" +
		"qgmtntw (82)\n" +
		"dmtlzbj (80)\n" +
		"zewvip (48)\n" +
		"znhkve (90)\n" +
		"odnfifx (80)\n" +
		"affmx (39)\n" +
		"pfeewo (638) -> fhhqd, szgqmow, bdeol, dyxlbp\n" +
		"rviwwhc (40)\n" +
		"wpqqze (84)\n" +
		"zmmfuq (10)\n" +
		"qpnyq (19)\n" +
		"kjvaqy (19)\n" +
		"dkuhniy (1187) -> nbenm, obnnecx, gqvxml, zyzwas\n" +
		"tzgsm (1080) -> jevck, uqsrhrf, orgcvcg\n" +
		"xrgsnsh (24) -> abwhrjo, xycwfmc, qqvlmd, omupor\n" +
		"ozolfes (85) -> fncqp, lusiwnm, anoxy\n" +
		"zzrttv (73)\n" +
		"dvdnlr (87)\n" +
		"qqubd (61)\n" +
		"adsppu (46)\n" +
		"rblolc (45) -> ooilg, fbcqhv, palqz, lqdvwk\n" +
		"blslvmm (381) -> tijjvh, oceimpw, lndzbn\n" +
		"qhdtqi (70)\n" +
		"obpwxg (7)\n" +
		"bxqlx (57) -> rqglm, tzlsvpk, gmsbmq, tdulat, cianrio\n" +
		"bqjqj (44)\n" +
		"uxfnv (153) -> brooyl, lwvclga, jeqxuvl\n" +
		"socihs (33)\n" +
		"qnioy (400)\n" +
		"mbcxw (19)\n" +
		"wytylnc (398) -> yinrfyl, kcewvb\n" +
		"hckgf (162) -> kutvex, prvai, jwdmmcd, safph, nyfutww, xochdg\n" +
		"mnsxmc (137) -> oadcp, adxgp\n" +
		"olyrohd (73)\n" +
		"nglln (12)\n" +
		"qoiit (67)\n" +
		"vqezl (64)\n" +
		"deboo (116) -> vmldb, vdujht\n" +
		"dystb (247) -> ryonf, ixqkcbm\n" +
		"dpqxwea (162) -> ccsqmpv, pjtzjkm\n" +
		"lpfprd (164) -> hcnocre, polzw\n" +
		"fviaonx (7)\n" +
		"brvlzkw (34)\n" +
		"ocpzwk (71) -> vwske, sdkbbc\n" +
		"ynhgsj (70) -> jzocbg, efylyde\n" +
		"orgcvcg (88) -> frqrk, wzinm\n" +
		"jfikyhd (74)\n" +
		"ajpjme (73) -> tymwxrt, kysjzj, qoiit\n" +
		"atquwvk (1437) -> vpqyy, hmxbpwl, mryloc, hxphn\n" +
		"csmkbgh (257) -> pqoybi, liyxgoa, zlrod, yvupc, lpfprd\n" +
		"hercuw (999) -> rpvqak, jynanod, gyzhdk, vonee\n" +
		"kpjmq (56)\n" +
		"iuosq (1130) -> zlyxnww, yiitzs\n" +
		"onxwvpl (47) -> nwuhqn, udpuj\n" +
		"ooilg (70)\n" +
		"dkolh (310) -> vfgkz, xsewp\n" +
		"crxcdnk (300) -> wkvcgtu, fnodwc\n" +
		"xycwfmc (99)\n" +
		"hxphn (12)\n" +
		"uhczfn (82)\n" +
		"juxbbz (52)\n" +
		"fqevwyy (92)\n" +
		"ybhtg (56) -> wbona, lwiyyiu, ieqvf, fultd\n" +
		"itaxno (12)\n" +
		"eibsqe (245) -> hojzxhu, ifmnxwy\n" +
		"lrzlckm (1087) -> hxoswpm, quomvk, xadrr, xmsfn, kyziqis\n" +
		"anlre (24)\n" +
		"anetug (70) -> gypqe, otise\n" +
		"syhbs (188) -> nglln, ogrcxm\n" +
		"wpmpzel (30)\n" +
		"rthbso (152) -> jlrun, jswlnwo, xziemca, vdkklj\n" +
		"narpa (248) -> yheewtp, tovvd\n" +
		"luralcy (55) -> dcaxloo, ovpoqt, bvrxeo, osgijzx, uppcjl\n" +
		"xurtoj (80)\n" +
		"hsycjdc (57)\n" +
		"geniumj (535) -> lieexcn, mchtb, nupqbfq\n" +
		"pcissqn (229) -> jplusbc, fiwwr\n" +
		"mebfx (74)\n" +
		"fhyhg (75)\n" +
		"pxsdzax (62)\n" +
		"waneo (24)\n" +
		"zwhpl (89)\n" +
		"vnlmjy (146) -> exhjhpu, jtwov\n" +
		"litns (1340) -> jbmccc, lefvu, anetug\n" +
		"qsdrdp (41) -> wdhjt, pqgnd, ftdjg, dzvsp\n" +
		"rxulpe (83)\n" +
		"vxpjfrf (106) -> eptfjt, zprpamt\n" +
		"slxrrx (62)\n" +
		"gwtsp (76)\n" +
		"hkrkkvv (55)\n" +
		"dbzmyy (97)\n" +
		"kywdy (99)\n" +
		"btsfhej (80)\n" +
		"acgdfu (70)\n" +
		"kccdhut (88)\n" +
		"rcfkr (70) -> khebz, oekex, gdhnu, uhggwqa, nrslon, incqze, ajpjme\n" +
		"jwdmmcd (172) -> odanj, nppone\n" +
		"phwnp (14)\n" +
		"asxxcu (13) -> vmhyd, ufzrbo, fkajvpp\n" +
		"jsylrrl (19)\n" +
		"qvbyk (96)\n" +
		"atazf (40)\n" +
		"fierzfm (10)\n" +
		"kwhjz (64)\n" +
		"vdkklj (23)\n" +
		"hnfwgag (54)\n" +
		"eptfjt (50)\n" +
		"sdxlyd (93) -> bqldhq, kpcorrf\n" +
		"azddjb (49)\n" +
		"ydktfd (88)\n" +
		"tgspt (43)\n" +
		"eqjoky (166) -> wwhao, wjncmeh\n" +
		"oadcp (18)\n" +
		"ycihhx (36)\n" +
		"bgacyt (66) -> fjforb, wsclc\n" +
		"takkclx (77)\n" +
		"jvhuwnd (50)\n" +
		"dmohbhf (56)\n" +
		"wdzqhs (102) -> kwnpdyr, yophj\n" +
		"fjforb (73)\n" +
		"npkqfq (12)\n" +
		"qugbhqd (35)\n" +
		"jxovdlo (6)\n" +
		"nmgme (75)\n" +
		"njmpeyi (17)\n" +
		"oekex (136) -> emjuzdq, twhkffp, onzxd\n" +
		"uewmev (32519) -> yvngku, kvgcvel, xupjwd\n" +
		"gmxfl (88)\n" +
		"fhjnytd (95)\n" +
		"ytoju (1052) -> tqveqn, vlzkx, gzcqdt, jwahk, xookni\n" +
		"lsyuq (87) -> vhdfs, jvhuwnd, vujcsg, oictp\n" +
		"taple (283)\n" +
		"fdojsjr (52)\n" +
		"sosqk (10)\n" +
		"rmvwkb (350) -> ffelpox, ftklgzk, fwxntdg\n" +
		"boifbq (1867) -> yscxlz, nijws, wnymiji\n" +
		"xdwuc (776) -> jhgxjnj, zdsyu, thufrr, gftjrqd, tgmtpht\n" +
		"xnhowa (70)\n" +
		"qxrnvvk (169) -> krvsaqc, uyvchvp\n" +
		"adxgp (18)\n" +
		"vsfhp (227) -> wnwzo, qslflr\n" +
		"jnyexah (834) -> dpqxwea, ckguaj, eqjoky, asxxcu\n" +
		"rzirj (24)\n" +
		"bqhfmb (11)\n" +
		"scwyfb (99) -> ctgjnch, nxbmvbe, mgklkfr, nozhmci\n" +
		"kwlqal (104) -> jhehhp, anhlx\n" +
		"aeppvjo (74) -> pouqokx, gmxfl\n" +
		"prwwpd (42)\n" +
		"zzuzfn (74)\n" +
		"rqgoz (201) -> jxovdlo, lmildeh\n" +
		"rgumam (96) -> fezzdc, vesqwkh, vzxaf, oztts\n" +
		"nouzec (15)\n" +
		"tqveqn (179)\n" +
		"bbpyckh (194) -> rrrrgbe, nyubyy\n" +
		"kpcorrf (51)\n" +
		"inhsin (26)\n" +
		"ixflbkx (812) -> afely, ikrsja, sjrmzwh, krrjt, zhclzz, iglop, bmnanm\n" +
		"crutxb (488) -> rblolc, iwvgg, kagubg, heqou, rhkgi\n" +
		"iacjli (48)\n" +
		"nescogt (38) -> oshhfb, ofkhwy, zdrtwa\n" +
		"jldua (19)\n" +
		"opttfsy (49)\n" +
		"vckfvm (100) -> svufvq, cpmuhnf\n" +
		"fodddp (73) -> znhkve, kbvepc\n" +
		"vbrhq (92)\n" +
		"wwhao (36)\n" +
		"zmnxzz (15)\n" +
		"eqhbool (269) -> rvfxkl, grosthd\n" +
		"gqejbh (44)\n" +
		"dzcfaed (40)\n" +
		"khebz (274)\n" +
		"qjhegw (84)\n" +
		"zmfwxre (72)\n" +
		"jlvppc (74) -> mlidg, xzvhic\n" +
		"furqyga (145) -> nssmfal, qaietz\n" +
		"qaietz (43)\n" +
		"lfbke (76)\n" +
		"fbbsb (24)\n" +
		"mxseaq (303)\n" +
		"oumbw (40)\n" +
		"vedrnbs (86)\n" +
		"zsfuc (671) -> hmgsla, chhab, csguji\n" +
		"ydzlkcn (47)\n" +
		"uppcjl (9128) -> nvgih, vjiqdn, nhrapmw\n" +
		"gxgvu (47)\n" +
		"wezzh (56)\n" +
		"npmifyp (54)\n" +
		"mloey (36) -> qvbyk, nqsgzg\n" +
		"oqihebu (51)\n" +
		"evnqnr (178)\n" +
		"nozhmci (69)\n" +
		"swojpec (21)\n" +
		"yjlokcq (49)\n" +
		"rmndp (18) -> srpftd, mebfx, olxrjth, iyeuk\n" +
		"ufzrbo (75)\n" +
		"hdhmvr (71)\n" +
		"hojzxhu (57)\n" +
		"eevkfzp (33)\n" +
		"lmildeh (6)\n" +
		"rsmfalh (43) -> ixeev, gwmkt\n" +
		"sleezka (36) -> sdfxsnj, weyts, jovejmr\n" +
		"cvwwx (77)\n" +
		"vwvpuxj (64)\n" +
		"rqvni (40)\n" +
		"yocxtug (922) -> ujvwff, tkmobkr, syhbs, bgacyt\n" +
		"ruayalk (89)\n" +
		"kbvepc (90)\n" +
		"codnc (8)\n" +
		"ovpoqt (55) -> ffwzbh, ntmbavi, lrzlckm, ptogh, boifbq, vdkwttr, dkuhniy\n" +
		"wbona (86)\n" +
		"vaedn (244)\n" +
		"pdrswn (65)\n" +
		"mqynznk (92)\n" +
		"dboryfe (263)\n" +
		"iwvgg (165) -> hxomy, xurtoj\n" +
		"kzuimi (110) -> yveusc, fkoaesc\n" +
		"vmldb (57)\n" +
		"kwnpdyr (74)\n" +
		"olxrjth (74)\n" +
		"cjvpndj (33)\n" +
		"zenhi (35)\n" +
		"dyxlbp (154)\n" +
		"hrbeox (13)\n" +
		"lieexcn (51) -> yyiiqr, uoyigbs, crqwarc, iacjli\n" +
		"xochdg (112) -> acgdfu, jrgndow\n" +
		"ffelpox (25) -> tplmdae, arowo\n" +
		"ukqajr (32)\n" +
		"tujzdkt (51)\n" +
		"pdvqv (72)\n" +
		"tftmhfs (75)\n" +
		"vjiqdn (1078) -> kuwbj, vckfvm, evnqnr\n" +
		"qyqir (79) -> vpfsdll, jmavf\n" +
		"brooyl (28)\n" +
		"ggkue (255) -> wtegiqv, gwrgur\n" +
		"zdrtwa (83) -> tezpif, txbdsy\n" +
		"wtenp (16)\n" +
		"xigac (52)\n" +
		"thgbcuu (65) -> fycbjfr, sdagro\n" +
		"aupbre (16)\n" +
		"eksxd (19)\n" +
		"gyzhdk (76) -> kdutnp, dhmnja, mfvta, jcbaauf\n" +
		"djjwlxc (99)\n" +
		"xacioc (56)\n" +
		"efylyde (80)\n" +
		"qomll (151) -> udydca, zhvyfkx\n" +
		"fycbjfr (60)\n" +
		"foqhhy (44)\n" +
		"znfjlz (94) -> lbvmy, ndhxim\n" +
		"djaamm (235) -> bxxpwa, pqjeof, itaxno\n" +
		"jxjrfer (213) -> olyrohd, iqyrubd\n" +
		"xbuysgv (6)\n" +
		"lkbpaz (124) -> wuoeas, ftudp\n" +
		"iysgr (586) -> lucoli, qnioy, ybhtg\n" +
		"ajncl (55)\n" +
		"mgklkfr (69)\n" +
		"yxjiuum (269) -> sokvf, squxpbv\n" +
		"jsyhr (472) -> swhkru, fkprqmd, ylwlpkw, uktdx, hhzarm\n" +
		"ocmqewc (7)\n" +
		"pddkiy (95)\n" +
		"bwifylg (32)\n" +
		"txrvdmv (65)\n" +
		"samoayn (10)\n" +
		"tijjvh (262) -> qwjcmlq, tggkndf\n" +
		"hgayc (155) -> zoviki, zewvip\n" +
		"sphbex (188)\n" +
		"sjrmzwh (26) -> qouck, ocezven, pvtnv\n" +
		"nyubyy (48)\n" +
		"xnyow (146) -> wfekg, opttfsy\n" +
		"gqrflb (99)\n" +
		"sudep (18)\n" +
		"rlngreu (93)\n" +
		"tymld (19)\n" +
		"kqieiv (77)\n" +
		"jyawoxq (73)\n" +
		"aeosriz (591) -> dhpgc, mgclfp, oivliv\n" +
		"nbdhcy (46)\n" +
		"tnbess (20)\n" +
		"rridhkb (84)\n" +
		"gzcqdt (67) -> dmohbhf, vgomsg\n" +
		"ckguaj (238)\n" +
		"cpxwsyb (13)\n" +
		"jfwbrfo (77)\n" +
		"mtvlfz (61)\n" +
		"ebptuar (9)\n" +
		"mfpxtom (128) -> xmwnu, rviwwhc\n" +
		"awxpkvm (75)\n" +
		"nupqbfq (227) -> crxfc, xcuzkj\n" +
		"craakno (119) -> ooiiqt, mfaftsr\n" +
		"jcbaauf (8) -> zvlise, takkclx, cvwwx, fnzocu\n" +
		"hvqdejm (202)\n" +
		"zftvgdy (170) -> swojpec, muuyp\n" +
		"rrrrgbe (48)\n" +
		"jswlnwo (23)\n" +
		"vfewf (77)\n" +
		"aqytxb (82)\n" +
		"njahbu (20)\n" +
		"jwahk (91) -> oxwpxqj, bqjqj\n" +
		"oztts (82)\n" +
		"ybijr (85)\n" +
		"zvernrt (104) -> vzdse, wrdaarz\n" +
		"oufye (39)\n" +
		"ltleg (808) -> vnlmjy, bscjk, uycjw, kzuimi, kwlqal, lahieha\n" +
		"cvewsdi (16)\n" +
		"yiitzs (62)\n" +
		"sverl (128) -> twxadm, iicum, efuepo\n" +
		"htipqs (55)\n" +
		"pbcywz (40) -> iwcqa, lhrdfkg\n" +
		"krmjp (123) -> avqmq, cstxcb\n" +
		"ltnnn (82)\n" +
		"eioql (1890) -> huutudx, azddjb\n" +
		"vwohcb (61)\n" +
		"aiiswrv (37) -> ydktfd, khxsoaa\n" +
		"srahmep (45)\n" +
		"yveusc (27)\n" +
		"exshg (10)\n" +
		"qgbdih (77)\n" +
		"gmhwcj (9)\n" +
		"nghznvw (31)\n" +
		"kewlet (44)\n" +
		"tvjrjqj (14)\n" +
		"pzths (18)\n" +
		"wojzq (34)\n" +
		"hcnocre (12)\n" +
		"jowbyue (37)\n" +
		"vwske (94)\n" +
		"khxsoaa (88)\n" +
		"nxbmvbe (69)\n" +
		"cydve (166) -> uruco, qryxc\n" +
		"wpyem (59) -> tnbjq, aexqgs, narpa\n" +
		"safph (102) -> eaggakx, qvyaen, dcbopj\n" +
		"ekchuez (76)\n" +
		"nsyqlu (56)\n" +
		"cosbycn (13)\n" +
		"xxylgr (46)\n" +
		"kurvox (73)\n" +
		"iicum (45)\n" +
		"rkcdvt (65) -> ozolfes, dystb, jsaksiq, taple\n" +
		"qfyzmmx (70)\n" +
		"xzhabw (69)\n" +
		"cnvgrhz (10)\n" +
		"emxhh (97) -> hnfwgag, ddzekr, rngyvds\n" +
		"zgcmic (1164) -> hvqdejm, nczcs, gpshsuk\n" +
		"nomegvq (31)\n" +
		"kvgcvel (6512) -> bnluwnh, eduzwi, ehjakg, atquwvk\n" +
		"gftkh (77)\n" +
		"odwmuu (16)\n" +
		"bqldhq (51)\n" +
		"awbgrbu (41) -> vuqpnz, xsnbmt\n" +
		"ccsqmpv (38)\n" +
		"vxzcqpq (4742) -> yciltsr, ohjniy, vrppcq, rleprim, anpriwx, furqyga, waunp\n" +
		"dcbopj (50)\n" +
		"tznxngl (90)\n" +
		"dhmnja (144) -> jgbgjbe, lwespc\n" +
		"awwywgr (145) -> lfijt, oqmjin\n" +
		"alvmkwx (47)\n" +
		"arowo (60)\n" +
		"knvukf (311) -> jnczao, xgkduuc, odwmuu\n" +
		"ffwzbh (90) -> jsthfix, vglnwmg, ggkue, khbylqn, xxwarr, qxrnvvk, djaamm\n" +
		"yygabb (84)\n" +
		"nssmfal (43)\n" +
		"klcixar (191) -> hkrkkvv, ezzht\n" +
		"szosumk (1834) -> ocpzwk, evlze, oqtjh\n" +
		"chhab (285) -> kzzzjj, swavc\n" +
		"incqze (274)\n" +
		"krrjt (101) -> pzprcdg, hgusv\n" +
		"xpxwgaq (36)\n" +
		"udydca (56)\n" +
		"lucdp (85) -> nebobeh, tccwm, vsfhp, okczka, klcixar, aenkx\n" +
		"ogrcxm (12)\n" +
		"tovvd (13)\n" +
		"btjjjxw (95) -> kwhjz, jmlnsj\n" +
		"utnhs (42)\n" +
		"xhzrq (8) -> eevkfzp, ajkkb\n" +
		"utoegrc (17)\n" +
		"gzttehr (33)\n" +
		"iypfs (49)\n" +
		"xihwd (52)\n" +
		"tkmobkr (62) -> eazsmg, tftmhfs\n" +
		"dtyaw (13)\n" +
		"yljnodb (62) -> iugub, xnyow, vaedn, gctjzc, rthbso, inkttc, zwaygv\n" +
		"mydomlh (96) -> sverl, qomll, dboryfe, ytiaiv, metlwn, vrlwo\n" +
		"nybkt (51)\n" +
		"jlrun (23)\n" +
		"ehjakg (726) -> axzjpa, jpuqyc, fodddp\n" +
		"yinrfyl (11)\n" +
		"qwdhug (197)\n" +
		"axklth (79)\n" +
		"lnwutyi (4943) -> islwhh, evjspos, lkbpaz, pbcywz, djoxie, wcyrbc\n" +
		"khcau (90) -> fkttvdm, vvbshe\n" +
		"xgqtx (184) -> kbrdsqg, bktqd\n" +
		"xgkduuc (16)\n" +
		"tpdaf (49) -> drukt, dvdnlr\n" +
		"wdhjt (44)\n" +
		"swakad (43)\n" +
		"ibvgkc (281)\n" +
		"fkttvdm (59)\n" +
		"ptogh (1126) -> yvidh, lsyuq, uvxcv\n" +
		"jeacolj (59)\n" +
		"xtcpdsl (36)\n" +
		"inkwv (92)\n" +
		"dqslnjk (31)\n" +
		"mqwndjo (1703) -> uojfba, lxpdska\n" +
		"vkyped (74)\n" +
		"pmhvbof (934) -> uysnnym, nhaxbnh, ufxskgv, thgbcuu\n" +
		"fhrui (105) -> eyhbm, srahmep\n" +
		"jpdvph (75)\n" +
		"sialf (19)\n" +
		"pnuluh (42)\n" +
		"tlcfxef (47)\n" +
		"niniqlk (12)\n" +
		"xxwarr (217) -> scpjvm, gwxeoes\n" +
		"ktcfyc (75)\n" +
		"qvyaen (50)\n" +
		"ssmfscq (42)\n" +
		"nfyjqfo (73)\n" +
		"rqglm (78) -> buoykr, fhyhg, wralkrd, efxmb\n" +
		"jznvixs (51)\n" +
		"uoyigbs (48)\n" +
		"bmatbfz (44) -> vqezl, yiuyo\n" +
		"zfavx (33) -> vpqmyx, bmaynas\n" +
		"exqsvn (16)\n" +
		"pbthb (64) -> abqvi, pfeewo, iuosq, nvpon, iujrwvw, zskbac, cgrcmg\n" +
		"ndkmiz (79)\n" +
		"wkvcgtu (7)\n" +
		"obnnecx (60) -> yuqqx, jpqcyh\n" +
		"ajkkb (33)\n" +
		"xookni (17) -> sziei, fnxtff\n" +
		"vgjtzi (55)\n" +
		"vdkwttr (947) -> xgqtx, hbvvpki, iprpx, wjboxd\n" +
		"kbrdsqg (38)\n" +
		"ixndjgk (100) -> surfn, njvmjgm\n" +
		"mgxfso (148) -> wxmyjrh, bwifylg\n" +
		"mexsmgu (40)\n" +
		"uemcawn (7)\n" +
		"rauprc (9)\n" +
		"zhclzz (179) -> affmx, nohqdkf\n" +
		"uhggwqa (19) -> iqbde, ybijr, zmizid\n" +
		"sjnpi (90)\n" +
		"wkjpw (50) -> ecconsw, vprox\n" +
		"oqmjin (26)\n" +
		"uysnnym (131) -> hygxqrs, memzmo\n" +
		"qpeztu (47)\n" +
		"tvuov (75)\n" +
		"xydikn (70)\n" +
		"ftklgzk (145)\n" +
		"fyynvhy (72)\n" +
		"odguqiy (46)\n" +
		"sccddm (86)\n" +
		"trvwm (47)\n" +
		"agsdkbw (30)\n" +
		"hxomy (80)\n" +
		"fwupt (74)\n" +
		"hbbkfas (40)\n" +
		"jsaksiq (117) -> fxbhth, wovwg\n" +
		"lmcqa (9)\n" +
		"hlxwud (42)\n" +
		"cgewz (64)\n" +
		"xtjals (13)\n" +
		"rpyji (1036) -> qbhev, feeksc, cjfnt, bkeqil, rkjfzl, nnqrj\n" +
		"dgsalqk (93)\n" +
		"islwhh (68) -> rridhkb, yygabb\n" +
		"ujvwff (94) -> uozuk, jeacolj\n" +
		"ieqvf (86)\n" +
		"cnsxofp (67)\n" +
		"crqwarc (48)\n" +
		"pzprcdg (78)\n" +
		"nvgih (1516) -> waneo, fesmk, qlzywpm, jdvsc\n" +
		"jgbgjbe (86)\n" +
		"rdqisk (68)\n" +
		"dzsjhs (1026) -> vvyizmq, iqdna, yvpxb\n" +
		"jnczao (16)\n" +
		"xiita (91) -> kwgmma, fhjnytd\n" +
		"ntmbavi (1078) -> mxseaq, fomagh, fdtmx\n" +
		"yrgxyez (108) -> ndczjq, cvewsdi, wtenp, exqsvn\n" +
		"ydrdiyq (7) -> peubkss, mricpy, bqwljb\n" +
		"fwlhy (20) -> vghgf, ecyxemp, crutxb\n" +
		"wqlrw (13)\n" +
		"iwkhee (25) -> vlnnmg, owsotek\n" +
		"bummi (9)\n" +
		"kcuhnx (116) -> djjwlxc, ikqttp\n" +
		"gcnhj (30)\n" +
		"gctjzc (64) -> hzkfzz, wwjqcpu\n" +
		"tplmdae (60)\n" +
		"cfzstl (86)\n" +
		"xvxrg (13)\n" +
		"bmxcqpu (6) -> ktcfyc, tvuov\n" +
		"sgjinm (49)\n" +
		"kimpf (145) -> tnbess, njahbu\n" +
		"jryikzp (84)\n" +
		"pjtzjkm (38)\n" +
		"lucoli (226) -> avgopx, lsjnlpy\n" +
		"gruqu (32)\n" +
		"vqglff (264) -> emxhh, yaxgb, tvzcg\n" +
		"ukixq (76)\n" +
		"wcqxiv (50)\n" +
		"quomvk (52) -> fwxkh, cgewz\n" +
		"pgcvs (65)\n" +
		"mchtb (153) -> jxpmi, hcfgsnr\n" +
		"wuoeas (56)\n" +
		"dmwzg (46)\n" +
		"hmgsla (145) -> odnfifx, btsfhej\n" +
		"efuepo (45)\n" +
		"zfkiukq (66)\n" +
		"ficpk (271) -> foqhhy, bdgldyi\n" +
		"pqqqof (39)\n" +
		"nnsaju (22)\n" +
		"ruqgy (306) -> oezxnl, kqxpxlx\n" +
		"fhhqd (60) -> alvmkwx, ydzlkcn\n" +
		"yvidh (127) -> kcitnaj, dmtlzbj\n" +
		"wnymiji (40)\n" +
		"hcfgsnr (45)\n" +
		"abqvi (480) -> zvernrt, obrdxor, wmnengd\n" +
		"uduzaf (32)\n" +
		"ytiaiv (122) -> tlcfxef, nrkxpd, gihgnqc\n" +
		"xoakdt (84)\n" +
		"tgmtpht (107) -> quaqp, tnyhegn, vjtyvg, llfxe\n" +
		"vzdse (77)\n" +
		"neisr (97) -> ukqajr, uduzaf, melui, gzxumn\n" +
		"tjfhwma (195)\n" +
		"lndzbn (48) -> rgnxext, xacioc, hhunwd, nsyqlu\n" +
		"zwaygv (90) -> gftkh, wtlmura\n" +
		"ndpsefd (24)\n" +
		"yvfkur (97)\n" +
		"uwcgpwx (10)\n" +
		"neshq (58)\n" +
		"axzjpa (153) -> kkgpo, wcqxiv\n" +
		"dhfiwjb (13)\n" +
		"sampa (10)\n" +
		"yvupc (86) -> jznvixs, tujzdkt\n" +
		"nhaxbnh (185)\n" +
		"mjeja (53)\n" +
		"iprpx (150) -> rhaqwc, pyulzo\n" +
		"wcyrbc (204) -> codnc, fcfhlc, vejds, rzpph\n" +
		"uqsrhrf (164) -> sulpwi, utoegrc\n" +
		"cvuaf (56)\n" +
		"eaugk (14)\n" +
		"xcuzkj (8)\n" +
		"pifqyxs (115) -> vwhmf, npmifyp\n" +
		"sbknkwh (82)\n" +
		"peubkss (75)\n" +
		"hakzty (81)\n" +
		"llgxywa (11)\n" +
		"qesfmt (74)\n" +
		"lsjnlpy (87)\n" +
		"efoouqi (104) -> bfejcbh, xihwd\n" +
		"jevck (156) -> sntwas, uijfe\n" +
		"buoykr (75)\n" +
		"xdpxpu (65117) -> zsfuc, litns, cslci\n" +
		"lwvclga (28)\n" +
		"yutvqn (225) -> xvxrg, zkjjvwh\n" +
		"jmavf (93)\n" +
		"cjfnt (737) -> nozlte, zvfxidm, sphbex\n" +
		"qunku (1662) -> xpxwgaq, krepomq, ycihhx\n" +
		"ikrsja (113) -> haxfzky, pdvqv\n" +
		"uojfba (94)\n" +
		"wfrqy (23) -> zwhpl, ozyvo, ruayalk\n" +
		"vedjm (51)\n" +
		"sdagro (60)\n" +
		"bktqd (38)\n" +
		"ooiiqt (34)\n" +
		"qijqu (88)\n" +
		"qtrappj (78)\n" +
		"jbbvph (369) -> avcuzkv, uqmdsne\n" +
		"vdujht (57)\n" +
		"owsotek (94)\n" +
		"qfcknc (97)\n" +
		"gbdida (46)\n" +
		"mjbmx (76)\n" +
		"uozuk (59)\n" +
		"cgemz (70)\n" +
		"ahjpc (43)\n" +
		"fmpkb (47)\n" +
		"jmdype (203) -> rgujeec, raezxi\n" +
		"sbgpr (287) -> fdyit, gruqu\n" +
		"ksphb (40)\n" +
		"mricpy (75)\n" +
		"peclf (84)\n" +
		"jeqxuvl (28)\n" +
		"fcoaxeb (628) -> uhgqfua, ddljunb, mgxfso\n" +
		"oshhfb (87) -> zfzli, hakzty\n" +
		"nnwxxk (64) -> rlngreu, dgsalqk\n" +
		"onzxd (46)\n" +
		"uijfe (21)\n" +
		"rlsynsa (63) -> qegjm, rogrsp\n" +
		"hfqskw (125) -> pqqqof, gleaooi, oufye\n" +
		"otise (6)\n" +
		"fdyit (32)\n" +
		"cptpfpd (65)\n" +
		"djoxie (52) -> vbrhq, akffhfg\n" +
		"xvkhbq (90)\n" +
		"gzxumn (32)\n" +
		"avqmq (31)\n" +
		"vyzfw (5391) -> hfqskw, wkjpw, gyifp, nqvbxx\n" +
		"lqdvwk (70)\n" +
		"ddzekr (54)\n" +
		"fdnmqri (29) -> afucrtw, gpium, pcissqn, hgayc, satrfh, rlsynsa, yutvqn\n" +
		"jioqtp (14)\n" +
		"cianrio (216) -> blhjjx, sdeecr\n" +
		"azlnl (73)\n" +
		"qnhgur (26)\n" +
		"mfvta (246) -> gxasczp, ayuwttz\n" +
		"njvmjgm (36)\n" +
		"utbrib (42)\n" +
		"mxhxyj (29)\n" +
		"yiuyo (64)\n" +
		"ayrojfl (46)\n" +
		"jakjzay (176) -> vrqga, dsugcog\n" +
		"fiwwr (11)\n" +
		"klnhysy (28)\n" +
		"zyzwas (174) -> xtjals, dtyaw\n" +
		"gleaooi (39)\n" +
		"hznrevv (77)\n" +
		"oceimpw (106) -> rxulpe, dtpzna\n" +
		"iqhrmo (63)\n" +
		"hhzarm (25) -> mexsmgu, bxqfge, oumbw\n" +
		"hygxqrs (27)\n" +
		"iqbde (85)\n" +
		"hgusv (78)\n" +
		"xrglp (958) -> odugb, ktbxbnn, vxpjfrf, xbegaua, nmtme\n" +
		"bmivh (103) -> gcnhj, wpmpzel, agsdkbw, quqdh\n" +
		"isgef (31)\n" +
		"dptakt (33)\n" +
		"wzwey (29)\n" +
		"yvngku (9809) -> fhsmlky, wdtrk, wpyem\n" +
		"gwmkt (77)\n" +
		"vrqga (18)\n" +
		"gcenos (90)\n" +
		"luzjos (161) -> wurmmqn, jioqtp, rpkaem, tvjrjqj\n" +
		"kcewvb (11)\n" +
		"pexst (46)\n" +
		"odanj (40)\n" +
		"nrslon (76) -> nczno, zfkiukq, yapxbpt\n" +
		"dmdgesa (92)\n" +
		"melui (32)\n" +
		"sokvf (41)\n" +
		"omupor (99)\n" +
		"tggkndf (5)\n" +
		"gtuoq (28)\n" +
		"czondc (13)\n" +
		"nrkxpd (47)\n" +
		"vwipyfi (44)\n" +
		"uhgqfua (106) -> tzvnssx, dqjxyyd\n" +
		"fvqwcn (868) -> gjdquua, pwaugpr, ncdavn\n" +
		"kmjzj (97)\n" +
		"zbicwki (7)\n" +
		"nrdcpk (44)\n" +
		"nnutid (21)\n" +
		"vesqwkh (82)\n" +
		"rhaqwc (55)\n" +
		"fyuwn (118) -> jsylrrl, qpnyq\n" +
		"wfekg (49)\n" +
		"mlidg (30)\n" +
		"vrfek (33) -> nlgewi, hlsch, szosumk, ixflbkx, ggmjg\n" +
		"tvzcg (171) -> nrdcpk, kewlet\n" +
		"segwmpm (255) -> fdojsjr, xigac\n" +
		"vllsfc (9)\n" +
		"donhvzg (77)\n" +
		"nczcs (170) -> noaiz, aupbre\n" +
		"pouqokx (88)\n" +
		"afucrtw (119) -> fajcjs, gsuxkc\n" +
		"nawvci (10)\n" +
		"temwow (75)\n" +
		"aceweuo (52)\n" +
		"fcfhlc (8)\n" +
		"gypqe (6)\n" +
		"ndczjq (16)\n" +
		"bmaynas (76)\n" +
		"kuwbj (68) -> vgjtzi, eovjzwr\n" +
		"ggsdtdp (12)\n" +
		"yaxgb (105) -> hznrevv, qgbdih\n" +
		"vxbdi (77)\n" +
		"gihgnqc (47)\n" +
		"fultd (86)\n" +
		"tezpif (83)\n" +
		"sudwjj (92)\n" +
		"fnihd (92)\n" +
		"dmneolr (6)\n" +
		"ighxxvd (88)\n" +
		"yshkxu (24)\n" +
		"vxaqay (23) -> qgmtntw, mjofjlf, sbknkwh, rwzme\n" +
		"lhrdfkg (98)\n" +
		"oarrrpk (22)\n" +
		"lxpdska (94)\n" +
		"yyiiqr (48)\n" +
		"sdsyt (225) -> ywtsd, iqhrmo\n" +
		"nkxpqbc (336) -> dukpbd, vwipyfi\n" +
		"weyts (40)\n" +
		"gmsbmq (378)\n" +
		"akffhfg (92)\n" +
		"odoffm (78) -> jpdvph, temwow\n" +
		"cgrcmg (1214) -> hnaps, sosqk, fierzfm, sampa\n" +
		"dyfyibk (76) -> cptpfpd, txrvdmv, pdrswn, ixfyyyz\n" +
		"awrwywl (4435) -> brjneo, nsqlbvb, vrfek, hxdrc, wfqul\n" +
		"jxpmi (45)\n" +
		"unxgbyg (41) -> azlnl, sqkae\n" +
		"lwespc (86)\n" +
		"udpuj (63)\n" +
		"cugwenm (82) -> qnhgur, inhsin\n" +
		"vzxaf (82)\n" +
		"fkajvpp (75)\n" +
		"vpqmyx (76)\n" +
		"mifct (46)\n" +
		"nozlte (188)\n" +
		"vfgkz (57)\n" +
		"rhkgi (257) -> wojzq, brvlzkw\n" +
		"ufxskgv (61) -> pxsdzax, cndzkwl\n" +
		"jmlnsj (64)\n" +
		"dcbsi (56)\n" +
		"pqyjyl (15)\n" +
		"fwxntdg (15) -> pgcvs, qekneh\n" +
		"mgclfp (77) -> ndkmiz, xvjuf, axklth, jgbce\n" +
		"zvfxidm (64) -> slxrrx, fiwcv\n" +
		"gynpo (175) -> eksxd, mbcxw\n" +
		"fwxkh (64)\n" +
		"tnbjq (252) -> qxtfl, llgxywa\n" +
		"nppone (40)\n" +
		"qxtfl (11)\n" +
		"fkprqmd (41) -> aceweuo, juxbbz\n" +
		"wiadoyw (285) -> kfmqmy, invtlwg\n" +
		"vyyye (70)\n" +
		"crxfc (8)\n" +
		"qrsyoj (33)\n" +
		"rgujeec (74)\n" +
		"bfejcbh (52)\n" +
		"oybkcy (33)\n" +
		"cslci (16) -> cydve, kcuhnx, crxcdnk, phkuv, rmndp\n" +
		"xsewp (57)\n" +
		"xupjwd (62) -> zgcmic, yocxtug, aeosriz, yljnodb, urwawsz, qunku, dzsjhs\n" +
		"uuziu (21)\n" +
		"twhkffp (46)\n" +
		"avcuzkv (20)\n" +
		"wjnoks (112) -> tzyexd, pzhayls\n" +
		"palqz (70)\n" +
		"hvnmuj (80) -> mifct, xxylgr\n" +
		"dmroo (197) -> itkyzhq, nomegvq, lszwjv\n" +
		"tptvyz (31)\n" +
		"uqmdsne (20)\n" +
		"ecconsw (96)\n" +
		"vejin (36)\n" +
		"czxsvq (30) -> gqrflb, kywdy\n" +
		"llfxe (29)\n" +
		"veokoh (687) -> yrgxyez, bmatbfz, hvnmuj, fclal, diauot, sgonpal, ixndjgk\n" +
		"urwawsz (1059) -> hoqzhm, awbgrbu, uxfnv\n" +
		"evjeyvt (563) -> ikwre, xhzrq, jfikyhd\n" +
		"kyziqis (180)\n" +
		"sdfxsnj (40)\n" +
		"yrrhqap (46)\n" +
		"tccwm (43) -> qswzv, iayecc, cfzstl\n" +
		"twxadm (45)\n" +
		"rzpph (8)\n" +
		"kzzzjj (10)\n" +
		"xzvhic (30)\n" +
		"qekneh (65)\n" +
		"cltcdp (170) -> tymld, jldua\n" +
		"vejds (8)\n" +
		"qzlzlwd (98)\n" +
		"ylwlpkw (115) -> utrhfs, cnvgrhz, exshg\n" +
		"qxfvj (93)\n" +
		"ywtsd (63)\n" +
		"hxoswpm (32) -> fwupt, qesfmt\n" +
		"ayocum (95)\n" +
		"vwueokp (96) -> xeoeht, cnsxofp\n" +
		"gwrgur (8)\n" +
		"cndzkwl (62)\n" +
		"diauot (100) -> rzirj, ndpsefd, atusqwd\n" +
		"cdogip (355) -> cosbycn, dhfiwjb, czondc\n" +
		"feeksc (29) -> dkolh, nkxpqbc, rgumam\n" +
		"rgnxext (56)\n" +
		"jpuqyc (22) -> ueefesu, donhvzg, prubmmc\n" +
		"vygkoys (18)\n" +
		"lruhxda (93)\n" +
		"bdeol (112) -> ckqlnc, uhiwblp, phwnp\n" +
		"ctldb (99)\n" +
		"ovyleaj (10)\n" +
		"fiwcv (62)\n" +
		"sulpwi (17)\n" +
		"dcaxloo (7979) -> blslvmm, pqozwy, rkcdvt, csmkbgh, jsyhr\n" +
		"ipllhv (148) -> sopkpcj, utbrib\n" +
		"vuqpnz (98)\n" +
		"emtol (26) -> oqihebu, gverkkt, dytqamx, xzkuda\n" +
		"jhehhp (30)\n" +
		"txbdsy (83)\n" +
		"sdeecr (81)\n" +
		"fnxtff (81)\n" +
		"mryloc (12)\n" +
		"nvpon (246) -> iagud, mdpooc, dyfyibk\n" +
		"iagud (256) -> atazf, ksphb\n" +
		"lpsufo (14)\n" +
		"qmczi (64)\n" +
		"dmraxdz (49)\n" +
		"ahnofa (7) -> xdpxpu, uewmev, awrwywl, hwezjo, qqqxyrl, luralcy\n" +
		"oyuzm (77)\n" +
		"jqsfxta (99)\n" +
		"kfmqmy (33)\n" +
		"ucapi (21)\n" +
		"eirbxu (92)\n" +
		"pmvbbt (50) -> hrgtmp, faszmg\n" +
		"phkuv (204) -> ajncl, htipqs\n" +
		"hzkfzz (90)\n" +
		"uycjw (66) -> dmraxdz, yjlokcq\n" +
		"pqozwy (345) -> aiiswrv, gynpo, iwkhee, rqgoz\n" +
		"mjjkr (32)\n" +
		"inkttc (205) -> hrbeox, wqlrw, cpxwsyb\n" +
		"oytgp (37)\n" +
		"bnluwnh (93) -> ipllhv, vpfbm, dftmlo, dbibp, akgkxvi, ydrdiyq\n" +
		"itkyzhq (31)\n" +
		"oivliv (363) -> nouzec, zmnxzz\n" +
		"lwiyyiu (86)\n" +
		"polzw (12)\n" +
		"hlsch (98) -> ficpk, nignkea, ramkjx, jxjrfer, untjkhr, mzqmah, eibsqe\n" +
		"yapxbpt (66)\n" +
		"vvyizmq (74) -> fcpeviq, kygcd, neshq\n" +
		"wdtrk (413) -> fyuwn, sleezka, pmvbbt\n" +
		"ilfgjk (76)\n" +
		"jplusbc (11)\n" +
		"tzvnssx (53)\n" +
		"ftdjg (44)\n" +
		"zlgdgn (522) -> onxwvpl, dqbrxb, mnsxmc\n" +
		"ayuwttz (35)\n" +
		"gftjrqd (211) -> xbuysgv, kzlpr\n" +
		"xdfuik (10)\n" +
		"npirsdf (84)\n" +
		"bqwljb (75)\n" +
		"fbcqhv (70)\n" +
		"wxmyjrh (32)\n" +
		"gpium (109) -> nwzzgtf, hdhmvr\n" +
		"tjjfh (25)\n" +
		"vpqyy (12)\n" +
		"gvqylt (56)\n" +
		"qjwrvpi (92)\n" +
		"crreuak (24)\n" +
		"oqtjh (15) -> pxjuos, qynhr, qqubd, lyzzn\n" +
		"frqrk (55)\n" +
		"fhzzvtv (37)\n" +
		"ycalpp (84)\n" +
		"paofmv (78)\n" +
		"fridz (31)\n" +
		"uyvchvp (51)\n" +
		"fncqp (66)\n" +
		"zynkd (70)\n" +
		"ixfyyyz (65)\n" +
		"tvbtw (25)\n" +
		"rbxfyau (12)\n" +
		"fclal (106) -> oybkcy, socihs\n" +
		"ggmjg (1823) -> ubcdd, rsmfalh, qwdhug, awwywgr\n" +
		"nnqrj (176) -> ozikd, fewkvyo, scwyfb\n" +
		"crxpsq (44)\n" +
		"gqvxml (200)\n" +
		"zkjjvwh (13)\n" +
		"evjspos (180) -> gtuoq, klnhysy\n" +
		"swavc (10)\n" +
		"gpshsuk (80) -> mtvlfz, wzoauoj\n" +
		"metlwn (233) -> samoayn, ovyleaj, uwcgpwx\n" +
		"wmnengd (104) -> vxbdi, kqieiv\n" +
		"zwgerw (44)\n" +
		"ygbuuxe (709) -> zfavx, krmjp, kimpf\n" +
		"cejbp (50)\n" +
		"hoqzhm (9) -> lfbke, ukixq, mjbmx\n" +
		"hhunwd (56)\n" +
		"efxdvtm (11)\n" +
		"vpfbm (64) -> peclf, qjhegw\n" +
		"kwgmma (95)\n" +
		"vlzkx (31) -> nxualql, vkyped\n" +
		"dbibp (160) -> xtcpdsl, iiozto\n" +
		"qqvlmd (99)\n" +
		"khbylqn (19) -> alvsbd, jryikzp, wpqqze\n" +
		"wjncmeh (36)\n" +
		"satrfh (133) -> mufdsrz, chdtaz\n" +
		"akmzxzw (46)\n" +
		"nqdcbhp (22)\n" +
		"ozkdsot (99)\n" +
		"efxmb (75)\n" +
		"qegjm (94)\n" +
		"ukmzkes (97)\n" +
		"uhiwblp (14)\n" +
		"sziei (81)\n" +
		"ncdavn (112) -> frqulg, xdfuik\n" +
		"xobzucu (22)\n" +
		"vglnwmg (251) -> zmmfuq, nawvci\n" +
		"fhgujr (317) -> ayrojfl, pexst\n" +
		"bfekpxz (359)\n" +
		"wtegiqv (8)\n" +
		"scpjvm (27)\n" +
		"obrdxor (74) -> sudwjj, eirbxu\n" +
		"vjtyvg (29)\n" +
		"ikqttp (99)\n" +
		"kysjzj (67)\n" +
		"thufrr (51) -> sccddm, orlxs\n" +
		"bdgldyi (44)\n" +
		"dffdie (786) -> nnwxxk, aeppvjo, wdzqhs, hogtrz\n" +
		"tnyhegn (29)\n" +
		"nmtme (22) -> pgodqz, qjwrvpi\n" +
		"gdiqocb (34) -> lxnip, vwohcb\n" +
		"nbenm (56) -> fyynvhy, zmfwxre\n" +
		"dqbrxb (110) -> lxckefh, lhvil, uuziu\n" +
		"qftepgn (70)\n" +
		"fhsmlky (38) -> nhjbg, xiita, eqhbool\n" +
		"osgijzx (7268) -> mydomlh, hckgf, tzgsm, pmhvbof\n" +
		"jfwcwqn (84)\n" +
		"ozyvo (89)\n" +
		"onjzq (85) -> zynkd, vyyye\n" +
		"erwjvd (1820) -> nqcfsr, kpjmq, jvjlfb\n" +
		"untjkhr (311) -> yshkxu, anlre\n" +
		"afely (164) -> isgef, vlucb, fridz\n" +
		"ctgjnch (69)\n" +
		"wrdaarz (77)\n" +
		"vgomsg (56)\n" +
		"bscjk (66) -> pplqbii, sgjinm\n" +
		"brjneo (1742) -> lucdp, lkhka, xdwuc, ixdllwp, veokoh, mqwndjo\n" +
		"squxpbv (41)\n" +
		"ezzht (55)\n" +
		"lahieha (20) -> hfcotwm, vejin, ohsjl, nsrpwww\n" +
		"wfqul (1160) -> erwjvd, eioql, rcfkr, xrglp, lyetyd, fjgzrim\n" +
		"jynanod (545) -> qyqir, vxfoyx, bjhhjla\n" +
		"msokpnb (7)\n" +
		"sgonpal (60) -> dcbsi, cvuaf\n" +
		"dytqamx (51)\n" +
		"xqscl (37)\n" +
		"fewkvyo (325) -> tvbtw, tjjfh\n" +
		"ecstxkl (78)\n" +
		"bxqfge (40)\n" +
		"vmhyd (75)\n" +
		"fomagh (11) -> zzrttv, nfyjqfo, jyawoxq, kurvox\n" +
		"orlxs (86)\n" +
		"viwzkp (40)\n" +
		"lbvmy (59)\n" +
		"avgopx (87)\n" +
		"abwhrjo (99)\n" +
		"qynhr (61)\n" +
		"lyzzn (61)\n" +
		"ozikd (333) -> hkckef, xorzt\n" +
		"pyulzo (55)\n" +
		"wjboxd (14) -> qlmptj, iyvjjg, kewmzq\n" +
		"fnzocu (77)\n" +
		"kcitnaj (80)\n" +
		"akgkxvi (148) -> prwwpd, bskqnj\n" +
		"vrppcq (51) -> gcenos, xvkhbq\n" +
		"njvbqvi (75)\n" +
		"zoviki (48)\n" +
		"vlucb (31)\n" +
		"jpqcyh (70)\n" +
		"dgjzv (46)\n" +
		"nqvbxx (224) -> wrmngqs, kcjcpk\n" +
		"sewnmby (40)\n" +
		"atusqwd (24)\n" +
		"odugb (142) -> mjjkr, qdqphk\n" +
		"zlyxnww (62)\n" +
		"wtlmura (77)\n" +
		"dftmlo (127) -> qugbhqd, pehkkzk, zenhi\n" +
		"lfijt (26)\n" +
		"lkqqjqh (76)\n" +
		"jzocbg (80)\n" +
		"csguji (8) -> kralmj, ozkdsot, ctldb\n" +
		"ixqkcbm (18)\n" +
		"nlgewi (1927) -> czxsvq, mloey, odoffm\n" +
		"wralkrd (75)\n" +
		"dketeda (574) -> vwueokp, emtol, wjnoks\n" +
		"kagubg (45) -> xydikn, qftepgn, qhdtqi, jbqnyve\n" +
		"kdutnp (304) -> dmneolr, acrhuro\n" +
		"jtwov (9)\n" +
		"tzyexd (59)\n" +
		"frqulg (10)\n" +
		"grlbob (92)\n" +
		"qmbfra (214) -> njmpeyi, hwnasq\n" +
		"fcpeviq (58)\n" +
		"qqqxyrl (31721) -> ogyypi, hercuw, lnwutyi, vxzcqpq, fwlhy, vyzfw\n" +
		"iujrwvw (1047) -> kyoypr, cxtxnm, xzhabw\n" +
		"wzoauoj (61)\n" +
		"fdtmx (135) -> kczko, npirsdf\n" +
		"sntwas (21)\n" +
		"oegamc (55) -> xnhowa, tfpsrke\n" +
		"cxtxnm (69)\n" +
		"hogtrz (214) -> niniqlk, ahbxz, npkqfq\n" +
		"hrgtmp (53)\n" +
		"vonee (215) -> vakapy, fbshzlu, neisr, xdqma, onjzq\n" +
		"nsrpwww (36)\n" +
		"ofkhwy (97) -> ilfgjk, lkqqjqh\n" +
		"ddljunb (58) -> oyuzm, jfwbrfo\n" +
		"jbpwoh (195)\n" +
		"fjgzrim (1337) -> luzjos, imruiet, qsdrdp\n" +
		"kkgpo (50)\n" +
		"ygono (78)\n" +
		"sopkpcj (42)\n" +
		"fxbhth (83)\n" +
		"rvfxkl (6)\n" +
		"dahvvo (23)\n" +
		"memzmo (27)\n" +
		"gfyvv (29) -> kmjzj, yvfkur\n" +
		"fnodwc (7)\n" +
		"alvsbd (84)\n" +
		"evlze (167) -> gbdida, dgjzv\n" +
		"ikves (54) -> ukmzkes, rkjdd\n" +
		"lhvil (21)\n" +
		"gxasczp (35)\n" +
		"jdvsc (24)\n" +
		"mimilpa (106) -> obpwxg, ocmqewc, zbicwki, msokpnb\n" +
		"nignkea (253) -> mvavmml, mjeja\n" +
		"rkjdd (97)\n" +
		"bskqnj (42)\n" +
		"anoxy (66)\n" +
		"noaiz (16)\n" +
		"yvpxb (206) -> ucapi, nnutid\n" +
		"cbcxlyo (50)\n" +
		"tbyfd (23)\n" +
		"aenkx (301)\n" +
		"pqgnd (44)\n" +
		"nsqlbvb (10733) -> evjeyvt, nescogt, rmvwkb\n" +
		"zkljkp (58) -> ayocum, pddkiy\n" +
		"wzinm (55)\n" +
		"bkeqil (119) -> qmoorx, cdogip, csiof\n" +
		"rngyvds (54)\n" +
		"jhgxjnj (223)\n" +
		"xmsfn (8) -> vedrnbs, gynzo\n" +
		"grosthd (6)\n" +
		"kieip (84)\n" +
		"qwcyeqz (33)\n" +
		"imruiet (37) -> tznxngl, sjnpi\n" +
		"pgodqz (92)\n" +
		"rmvhy (99)\n" +
		"eqnowu (31)\n" +
		"ramkjx (337) -> bqhfmb, efxdvtm\n" +
		"ndhxim (59)\n" +
		"jwyhe (582) -> tjfhwma, nshgykp, sdxlyd, vtnfaa, oegamc, jbpwoh, fhrui\n" +
		"bjhhjla (171) -> fmpkb, qpeztu\n" +
		"nyfutww (54) -> rmvhy, jqsfxta\n" +
		"gjdquua (52) -> rqvni, viwzkp\n" +
		"huutudx (49)\n" +
		"ueefesu (77)\n" +
		"invtlwg (33)\n" +
		"oyvsrsc (106) -> inkwv, fnihd\n" +
		"lxnip (61)\n" +
		"prvai (182) -> stvon, hhumkf\n" +
		"ogyypi (3236) -> sbgug, zlgdgn, vqglff\n" +
		"vujcsg (50)\n" +
		"dhktab (56) -> ygono, qtrappj\n" +
		"ckqlnc (14)\n" +
		"jrovat (168) -> tptvyz, eqnowu\n" +
		"mjofjlf (82)\n" +
		"vghgf (7) -> wiadoyw, sbgpr, jmdype, sdsyt, yxjiuum, vxaqay\n" +
		"lszwjv (31)\n" +
		"hbvvpki (166) -> gxgvu, trvwm\n" +
		"dzvsp (44)\n" +
		"pplqbii (49)\n" +
		"ovpmj (82)\n" +
		"bgwlsu (92)\n" +
		"yrudth (15)\n" +
		"txrejp (49)\n" +
		"wsclc (73)\n" +
		"oezxnl (57)\n" +
		"swhkru (43) -> nybkt, vedjm\n" +
		"kralmj (99)\n" +
		"girvl (43)\n" +
		"yciltsr (73) -> xyoatzj, bcmep\n" +
		"quqdh (30)\n" +
		"qmoorx (58) -> lwuwrp, czrmixj, ycalpp, jfwcwqn\n" +
		"hnaps (10)\n" +
		"gdhnu (194) -> hbbkfas, dzcfaed\n" +
		"xadrr (92) -> gqejbh, zwgerw\n" +
		"gyifp (114) -> vwvpuxj, qmczi\n" +
		"ecyxemp (1270) -> ibvgkc, ericb, zeitsku\n" +
		"anhlx (30)\n" +
		"wxyel (49)\n" +
		"svufvq (39)\n" +
		"lkhka (1423) -> aemfgyt, gdiqocb, bmxcqpu\n" +
		"acrhuro (6)\n" +
		"uriolfn (44)\n" +
		"nczno (66)\n" +
		"fezzdc (82)\n" +
		"mzqmah (83) -> mqynznk, fqevwyy, cbldohy\n" +
		"ftmwpqg (39) -> zzuzfn, gwhfv\n" +
		"ozmbhy (862) -> cugwenm, mimilpa, jlvppc\n" +
		"eiisyk (31)\n" +
		"stvon (35)\n" +
		"pehkkzk (35)\n" +
		"lusiwnm (66)\n" +
		"raezxi (74)\n" +
		"ericb (127) -> xogkc, vfewf\n" +
		"dhpgc (393)\n" +
		"wxloqgs (223)\n" +
		"hwnasq (17)\n" +
		"fkoaesc (27)\n" +
		"zfzli (81)\n" +
		"gewtd (59) -> bojkbqr, njvbqvi, nmgme, awxpkvm\n" +
		"nqcfsr (56)\n" +
		"lvmyjp (33)\n" +
		"waunp (119) -> wezzh, gvqylt\n" +
		"xmwnu (40)\n" +
		"uvxcv (91) -> txrejp, wxyel, iypfs, wbbfc\n" +
		"liyxgoa (130) -> mxhxyj, wzwey\n" +
		"okczka (105) -> mmsesaw, qzlzlwd\n" +
		"qdqphk (32)\n" +
		"jbmccc (34) -> fbbsb, crreuak\n" +
		"wprzdkf (37) -> jbbvph, fhgujr, zcubrms\n" +
		"fajcjs (66)\n" +
		"dsugcog (18)\n" +
		"xogkc (77)\n" +
		"mufdsrz (59)\n" +
		"ubcdd (197)\n" +
		"nqsgzg (96)\n" +
		"nshgykp (129) -> lvmyjp, cjvpndj\n" +
		"qlzywpm (24)\n" +
		"kewmzq (82)\n" +
		"nijws (40)\n" +
		"prubmmc (77)\n" +
		"mdpooc (250) -> swakad, girvl\n" +
		"rwzme (82)\n" +
		"bmnanm (173) -> pnuluh, hlxwud\n" +
		"vprox (96)\n" +
		"yuqqx (70)\n" +
		"iglop (89) -> xoakdt, kieip\n" +
		"pwaugpr (6) -> utnhs, xjqta, ssmfscq\n" +
		"iyvjjg (82)\n" +
		"zvlise (77)\n" +
		"vtnfaa (43) -> ekchuez, gwtsp\n" +
		"iiozto (36)\n" +
		"hhumkf (35)\n" +
		"etzfe (18)\n" +
		"bsmwi (57)\n" +
		"xzkuda (51)\n" +
		"zprpamt (50)\n" +
		"zeitsku (71) -> qfyzmmx, cgemz, kqaox\n" +
		"vvbshe (59)\n" +
		"nxualql (74)\n" +
		"srpftd (74)\n" +
		"gwxeoes (27)\n" +
		"gynzo (86)\n" +
		"wnwzo (37)";

	@Test
	public void testPart1() throws Exception {
		assertEquals("tknk", new Day07().part1("pbga (66)\n" +
			"xhth (57)\n" +
			"ebii (61)\n" +
			"havc (66)\n" +
			"ktlj (57)\n" +
			"fwft (72) -> ktlj, cntj, xhth\n" +
			"qoyq (66)\n" +
			"padx (45) -> pbga, havc, qoyq\n" +
			"tknk (41) -> ugml, padx, fwft\n" +
			"jptl (61)\n" +
			"ugml (68) -> gyxo, ebii, jptl\n" +
			"gyxo (61)\n" +
			"cntj (57)\n"));
	}

	@Test
	public void testPart2() throws Exception {
		assertEquals(60, new Day07().part2("pbga (66)\n" +
			"xhth (57)\n" +
			"ebii (61)\n" +
			"havc (66)\n" +
			"ktlj (57)\n" +
			"fwft (72) -> ktlj, cntj, xhth\n" +
			"qoyq (66)\n" +
			"padx (45) -> pbga, havc, qoyq\n" +
			"tknk (41) -> ugml, padx, fwft\n" +
			"jptl (61)\n" +
			"ugml (68) -> gyxo, ebii, jptl\n" +
			"gyxo (61)\n" +
			"cntj (57)\n"));
	}

	@Test
	public void actualPart1() throws Exception {
		System.out.println(new Day07().part1(INPUT));
	}

	@Test
	public void actualPart2() throws Exception {
		System.out.println(new Day07().part2(INPUT));
	}


}
